0-print_listint.c

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,14 +1,13 @@
 #include "lists.h"
 
 /**
- * print_listint - prints all the elements of a linked list
+ * print_listint - prints all elements of a linked list
  * @h: linked list of type listint_t to print
- *
- * Return: number of nodes
+ * Return: The No of nodes
  */
 size_t print_listint(const listint_t *h)
 {
-	size_t j = 0;
+	int j = 0;
 
 	if (h != NULL)
 	{
